Remove unused variable check on numbers
In order to detect that we have to highlight numbers we reuse the
regexpes from the Python syntax file.
However, since I cleaned up some variables in the file, I forgot to
remove the condition check. This results in numbers not being
highlighted at all.

This patch removes the check on a no-longer existing variable.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>