Remove criteria value higlighting
Some specific values such as True,False and none were highlighting as
CriteriaValues. This makes sense if every criteria value could be
highlighting.
However, those values are user-defined and not known in advance.
It has no sense to partially highlight them, so the support should be
removed.

This patch removes the support for partial CriteriaValues highlight for
a more consistent behavior.
It fixes #3.

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>